Moved starting files to src and added background
added a background - need to dim the background later

Also added code for better structural code - MAKE SURE TO UPDATE UR LOCAL CODE TO THIS ONE
