+++ src/code/structure/setting.py
@@ -1,0 +1,5 @@
+class Settings:
+    def __init__(self):
+        self.screen_width = 1280
+        self.screen_height = 720
+
