+++ src/data/Tilesets/House Decoration.tsx
