+++ src/data/Tilesets/Paths.tsx
