+++ src/data/Tilesets/Plant Decoration.tsx
