+++ src/graphics/osu_logo.png
